requirements.txt
anyio==3.5.0
asgiref==3.5.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
ecdsa==0.17.0
fastapi==0.76.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
iniconfig==1.1.1
outcome==1.1.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
starlette==0.18.0
tomli==2.0.1
trio==0.20.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
